requirements.txt
telethon
requests
pyMorseTranslator